+++ src/yw2htmllib/html_export.py
@@ -11,7 +11,7 @@ from pywriter.file.file_export import FileExport
 class HtmlExport(FileExport):
     """HTML file representation.
     
-    Provide basid HTML templates for exporting chapters and scenes.
+    Provide basic HTML templates for exporting chapters and scenes.
     """
     DESCRIPTION = 'HTML export'
     EXTENSION = '.html'
@@ -61,93 +61,6 @@ strong {font-weight:normal; text-transform: uppercase}
 </html>
 '''
 
-    def _get_chapterMapping(self, chId, chapterNumber):
-        """Return a mapping dictionary for a chapter section. 
-
-        Positional arguments:
-            chId -- str: chapter ID.
-            chapterNumber -- int: chapter number.
-
-        Extends the superclass method.
-        """
-        ROMAN = [
-            (1000, "M"),
-            (900, "CM"),
-            (500, "D"),
-            (400, "CD"),
-            (100, "C"),
-            (90, "XC"),
-            (50, "L"),
-            (40, "XL"),
-            (10, "X"),
-            (9, "IX"),
-            (5, "V"),
-            (4, "IV"),
-            (1, "I"),
-        ]
-
-        def number_to_roman(n):
-            """Return n as a Roman number.
-            
-            Credit goes to the user 'Aristide' on stack overflow.
-            https://stackoverflow.com/a/47713392
-            """
-            result = []
-            for (arabic, roman) in ROMAN:
-                (factor, n) = divmod(n, arabic)
-                result.append(roman * factor)
-                if n == 0:
-                    break
-
-            return "".join(result)
-
-        TENS = {30: 'thirty', 40: 'forty', 50: 'fifty',
-                60: 'sixty', 70: 'seventy', 80: 'eighty', 90: 'ninety'}
-        ZERO_TO_TWENTY = (
-            'zero', 'one', 'two', 'three', 'four', 'five', 'six', 'seven', 'eight', 'nine', 'ten',
-            'eleven', 'twelve', 'thirteen', 'fourteen', 'fifteen', 'sixteen', 'seventeen', 'eighteen', 'nineteen', 'twenty'
-        )
-
-        def number_to_english(n):
-            """Return n as a number written out in English.
-
-            Credit goes to the user 'Hunter_71' on stack overflow.
-            https://stackoverflow.com/a/51849443
-            """
-            if any(not x.isdigit() for x in str(n)):
-                return ''
-
-            if n <= 20:
-                return ZERO_TO_TWENTY[n]
-
-            elif n < 100 and n % 10 == 0:
-                return TENS[n]
-
-            elif n < 100:
-                return f'{number_to_english(n - (n % 10))} {number_to_english(n % 10)}'
-
-            elif n < 1000 and n % 100 == 0:
-                return f'{number_to_english(n / 100)} hundred'
-
-            elif n < 1000:
-                return f'{number_to_english(n / 100)} hundred {number_to_english(n % 100)}'
-
-            elif n < 1000000:
-                return f'{number_to_english(n / 1000)} thousand {number_to_english(n % 1000)}'
-
-            return ''
-
-        chapterMapping = super()._get_chapterMapping(chId, chapterNumber)
-        if chapterNumber:
-            chapterMapping['ChNumberEnglish'] = number_to_english(chapterNumber).capitalize()
-            chapterMapping['ChNumberRoman'] = number_to_roman(chapterNumber)
-        else:
-            chapterMapping['ChNumberEnglish'] = ''
-            chapterMapping['ChNumberRoman'] = ''
-        if self.chapters[chId].suppressChapterTitle:
-            chapterMapping['Title'] = ''
-        return chapterMapping
-
     def _convert_from_yw(self, text, quick=False):
         """Return text, converted from yw7 markup to HTML markup.
         
@@ -165,24 +78,30 @@ strong {font-weight:normal; text-transform: uppercase}
                 return ''
             else:
                 return text
-            
-        HTML_REPLACEMENTS = [
-            ('\n', '</p>\n<p>'),
-            ('[i]', '<em>'),
-            ('[/i]', '</em>'),
-            ('[b]', '<strong>'),
-            ('[/b]', '</strong>'),
-            ('<p></p>', '<p><br /></p>'),
-            ('/*', '<!--'),
-            ('*/', '-->'),
-        ]
-        try:
+
+        if text:
+            # Remove inline code.
+            YW_SPECIAL_CODES = ('HTM', 'TEX', 'RTF', 'epub', 'mobi', 'rtfimg', 'RTFBRK')
+            for specialCode in YW_SPECIAL_CODES:
+                text = re.sub(f'\<{specialCode} .+?\/{specialCode}\>', '', text)
+
+            # Apply html formatting.
+            HTML_REPLACEMENTS = [
+                ('\n', '</p>\n<p>'),
+                ('[i]', '<em>'),
+                ('[/i]', '</em>'),
+                ('[b]', '<strong>'),
+                ('[/b]', '</strong>'),
+                ('<p></p>', '<p><br /></p>'),
+                ('/*', '<!--'),
+                ('*/', '-->'),
+            ]
             for yw, htm in HTML_REPLACEMENTS:
                 text = text.replace(yw, htm)
 
             # Remove highlighting, alignment,
             # strikethrough, and underline tags.
             text = re.sub('\[\/*[h|c|r|s|u]\d*\]', '', text)
-        except AttributeError:
+        else:
             text = ''
         return(text)